+++ src/utils.hpp
@@ -99,9 +99,9 @@ namespace BI
 
 	namespace cocos
 	{
+#ifdef GEODE_IS_WINDOWS
 		inline cocos2d::CCPoint getMousePosition()
 		{
-#ifdef GEODE_IS_WINDOWS
 			auto* director = cocos2d::CCDirector::sharedDirector();
 			auto* gl = director->getOpenGLView();
 			auto winSize = director->getWinSize();
@@ -109,10 +109,8 @@ namespace BI
 			auto mouse = gl->getMousePosition() / frameSize;
 
 			return cocos2d::CCPoint{ mouse.x, 1.f - mouse.y } * winSize;
-#elif defined(GEODE_IS_MACOS)
-			return geode::cocos::getMousePos();
-#endif
 		}
+#endif
 
 		inline bool isPositionInNode(cocos2d::CCNode* node, const cocos2d::CCPoint& pos)
 		{
@@ -132,8 +130,12 @@ namespace BI
 
 	enum class PlatformKey
 	{
+		// Windows Control key | MacOS Command key
 		LEFT_CONTROL,
-		LEFT_SHIFT
+		// Shift key
+		LEFT_SHIFT,
+		// Window Alt key | MacOS Option key
+		LEFT_ALT
 	};
 	namespace platform
 	{
@@ -146,6 +148,8 @@ namespace BI
 					return GetKeyState(VK_CONTROL) & 0x8000;
 				case BI::PlatformKey::LEFT_SHIFT:
 					return GetKeyState(VK_SHIFT) & 0x8000;
+				case BI::PlatformKey::LEFT_ALT:
+					return GetKeyState(VK_LMENU) & 0x8000;
 			}
 
 			return false;